integrated migrations & in-memory keystore
