+++ docker-compose.yml
@@ -10,6 +10,7 @@ services:
       PASSAGE_HTTP_ADDR: 0.0.0.0:6000
       PASSAGE_API_ENABLED: 'true'
       PASSAGE_TUNNEL_STANDARD_ENABLED: 'true'
+      PASSAGE_KEYSTORE_POSTGRES_TABLE_NAME: 'keystore.keys'
 
       PGHOST: postgres
       PGPORT: 5432
@@ -20,6 +21,8 @@ services:
       - "6000:6000"
     networks:
       - passage_dev
+    volumes:
+      - ./sql:/docker-entrypoint-initdb.d
     depends_on:
       postgres:
         condition: service_healthy
