+++ sql/keystore.sql
@@ -1,5 +1,7 @@
+CREATE SCHEMA IF NOT EXISTS keystore;
+
 CREATE EXTENSION IF NOT EXISTS "uuid-ossp";
-CREATE TABLE IF NOT EXISTS passage.keys(
+CREATE TABLE IF NOT EXISTS keystore.keys(
    id           UUID DEFAULT uuid_generate_v4(),
    created_at   timestamp without time zone DEFAULT CURRENT_TIMESTAMP NOT NULL,
    contents     VARCHAR,
