+++ test/docker-compose.yml
@@ -13,8 +13,6 @@ services:
       PASSAGE_TUNNEL_REVERSE_ENABLED: 'true'
 
       PASSAGE_DISCOVERY_STATIC_HOST: 'passage'
-      PASSAGE_KEYSTORE_TYPE: 'postgres'
-      PASSAGE_KEYSTORE_POSTGRES_TABLE_NAME: 'passage.keys'
       LOG_LEVEL: debug
 
       PGHOST: postgres
